Factor database into own module
